feat(Composer): add composer to manage dependencies and autoload
