--- dao/MemberDao.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace testAnne\blog\dao;
-
-use testAnne\blog\model\member;
-
-class MemberDao extends BaseDao
-{
-
-}
