+++ script/import_script.sql
@@ -1,0 +1,13 @@
+/**
+ * CREATE DATABASE
+ */
+DROP DATABASE IF EXISTS blog;
+CREATE DATABASE blog;
+USE blog;
+
+/**
+ * CREATE TABLES AND IMPORT DATA
+ */
+source blog.sql;
+
+
