+++ src/dao/MemberDao.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Blog\Dao;
+
+class MemberDao extends BaseDao
+{
+
+}
